requirements_render.txt
# ZamanYönet Render.com Requirements (Minimal - No Compilation)
# Pure Python packages only - guaranteed no C compilation

# Core Flask
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.20

# PostgreSQL (Pure Python)
pg8000==1.30.2

# Basic security (avoid cryptography & bcrypt for now)
PyJWT==2.8.0
flask-limiter==3.5.0
marshmallow==3.20.2

# Logging 
structlog==25.4.0

# Email support
flask-mail==0.9.1

# Validation
cerberus==1.3.5

# Production server
gunicorn==21.2.0

# Minimal dependencies only
python-decouple==3.8 